Move read_database functions to database_utilities.
